Fixed an MPI issue where "orted" was not found when spreading the processes out on multiple nodes. Fix is in scripts/pfcp.
Added logic so that sizes can be specified by using units, rather than exact numbers on
command line (and in the config file, I think). For example 1MB == 1048576 and 1mb == 1000000.

Added a synthetic data generation feature. Currently in order to enable it, configure needs
to be run with "--enable-syndata" set. Synthetic data is generated in the destination file
when the "-x <size>" is specified on the pfcp command line. A pattern, or a pattern file
can be specified as well with "-X <pattern file>" or "-X <pattern>". A file (or files) of
all zeros can be generated by specifying the special pattern "-X zero".