Changes to support parallel runs with MarFS destinations.
-- support for specifying expected-object-size in MarFS_Path::open()

there is new aupport in libaws4c to allow producing a content-length HTTP
header in the PUT request.  This makes a big difference in performance to
Scality (sproxyd), in curl command-line tests.  However, libcurl 7.19.7
seems not to support this correctly.  We've got the infrastructure in place
in libmarfs, by supporting a length parameter to marfe_open().  Everything
in place, here, to use that, once we figure out how to get libcurl to honor
it.

-- dusted off synthetic-input data facility.

It looked like this might be an easy way to get cheap low-cost input, in
the case where we just want to test write bandwidth.  It needed some
dusting off, but now looks like it will need a little more work to be used
for this particular problem.  It wants to create buffers witha fixed
pattern, which can be copied into the send-buffer, instead of reading from
a source.  We'd want to short-circuit that, to skip the copy.  TBD.

-- minor compile-time errors/warnings

-- cleaned up some of the hardwired path-manipulations in MARFS_Path

Instead of assuming paths start with "/marfs/", use marfs_sub_path() from
libmarfs, and use the info in marfs_config.

-- MARFS_Path learns the local MPI rank from the factory, at build-time

We'll be using this to allow pftool to target different hosts in a repo, so
that all the traffic doesn't go to the single host specified in the
configuration.  TBD.