Use PUT requests with explicit content-length headers.
The aws4c support for explicit content-length headers is now working, so we
engage it.  This means we're writing PUT requests that have content-length
matching the st.st_size of the source file.  THIS WILL BE WRONG for Multi
files.  In other words, this change breaks pftool, for writing Multi files.
The expected fix will be to compute the per-object write-size, and write
each MarFS object with its own correct length.

content-length vs. chunked-transfer-encoding seemed to make a big
difference in performance in some command-line tests, but doesn't seem to
produce an obvious improvement in parallel pftool runs.