Use new support in libmarfs, plus local exensions, to allow writing N:1 to MarFS
MarFS has some special-needs, that we wove into the Path type-hierarchy.
We added virtual methods Path::pre_process() and Path::post_process(),
which are called single-threaded (i.e. once) for each file that is being
copied.  These are called before opening, and after closing, respectively.
Pre-process gives MarFS an opportunity to identify the appropriate
repository to use, based on the full-size of the source-file.  Similarly,
post-process allows MarFS to truncate the MD file to the final size.  (We
don't do this in pre-process, because we plan to use the destination
file-size as a signal to users of the progress of a long parallel copy.

In the case of N:1 writes (i.e. chunking), the chunksize provided on the
command-line/configuration is ignored in favor of the chunksize configured
in marfs for the destination namespace, and the repo in that namespace
corresponding with the source file-size.

MarFS also requires each open() to declare the full-size of the
object-stream it is going to write, and to provide a logical offset
(i.e. we can do a "seek" after opening).  This length may actually be
greater than a MarFS chunk, but after writing zero or more chunks, the
final part of the length must either land at an object-boundary, or
represent the final chunk (in the whole file) in a MarFS "Multi" file.  To
support this, we now do all opens through a new Path::open() interface,
which has arguments that include an offset and a length.

TBD: MarFS metadata has the potential to support discovering how much of a
chunked file was actually written (via count_chunkinfo(), in libmarfs).
However, David Sherrill's recently-added restart techniques seem to "just
work" on top of MarFS.  We should keep an eye on any scaling issues, and
consider using the extensibility features in David's tech, to assimilate
count_chunkinfo() into the restart support, if we need to reduce the size
of xattrs or files that are the basis of the restart info.

For debugging, you can now use '-vvv' to get MarFS interactions to print
out all the curl headers and responses, and you can use '-vvvv' to get the
program to enter an infinite loop, waiting for the variable gdb to acquire
a value of 1.  This allows you to gdb attach to a worker (which waiting for
the master to enter a barrier), then gdb attach to the master, and set
gdb=1, and detach.