Add Path::open() interface with more parms, plus performance tweak
MarFS paths support the POSIX open/read/close semantics, but the actual GET
request submitted to object-storage backends benefits from knowing the
total length of the request, which is something availaable to pftool at
open-time.

So, we added a new Path::open() interface that allows us to telgraph the
total amount of data we expect to be reading, which can be tracked in the
underlying code, to add appropriate byte-ranges

Also, there were some calls to memset() that seemed unnecessary.  Wiping
the contents of a buffer before receiving data, for every transfer.
valigrind showed this taking a lot of time, so I commented it out.  There
was one in copy_file, and one in compare_file.