Adding experimental code to perform retries, when reads fail in copy_file().
When reading intensively from MarFS, it happens fairly often (~5-20% of
requests) that servers simply stop providing data for some of the requests,
or else there are long delays, even when the network has become quiet.
This causes timeouts resulting in read-failures.

As a first cut at dealing with this, we simply retry an entire block,
whenever a read fails to return the full size.  Furthermore, we do a close
and re-open.  Then reads proceed from that point as if nothing had
happened.  This solves our problem, but would be overkill for POSIX-based
systems, where there would be no need for the close/reopen.  For us, the
close-reopen submits a new GET request, which seems to be enough to get the
server to start talking to us again.

This solution could potentially be integrated more deeply into the MarFS
support, so that we'd never see it in pftool, and reads would internally do
the close/reopen, if they wanted to.  TBD.