Threaded MPI nprocs into PathFactory inits.  Unused, for now.
The idea was that this might eventually be needed to support extending the
"host-randomization" in the MarFS internals, for dealing with incast
problems.  For example, all tasks on a given host might connect to a
specific server IP, used only by them, and then fail-over to a different
one (also used only by them), if the first server stopped responding.  Or
they could pick servers at random from a subrange, used only by them.

This approach is probably not going to be used, as we now have some
per-host hashing in MarFS, as well as the discovery that breaking large
requests up into smaller ones seems to be reducing our incast problems.
But, I might as well commit this, so we'll have it later, if we need it.