"/dev/null/" destination allows no-cost writes, which simplify read-BW tests.
We want to give the appearance of a sort of "/dev/null directory tree", to
allow pftool to show read BW in the case where the only available
destination file-system is slow (GPFS), and there is insufficient tempfs
space to allow that to be used instead.  In this case, pftool will believe
it is creating destination directories, writing files, etc, but all these
operations are no-ops.

The result is that read-performance is not constrained by slow
write-performance.

To use this, the destination (-c) should be "/dev/null", or "/dev/null/".
It makes a difference whether you add the final slash, or not.  "/dev/null"
will be treated as a file, whereas "/dev/null/" looks like a directory.

For the file case, stat() returns a stat of 'dev/null'

For the directory case, stat() does a stat of the user's homedir, mkdir()
always succeeds (without doing anything), and readdir() always shows an
empty directory.  No operations will touch anything in the user's homedir,
we just use it to make a realistic directory stat struct.