Fix updates to destination ctime/atime, so N:N restarts work again.
MARFS_Path::chown(),chmod(),utime() were assuming they could just mark the
status as obsolete, after making these changes to the underlying file, and
the stat would thus be reperformed when needed.  That doesn't work if the
Path goes out of scope.  Factory will see the old stat and assume it is
up-to-date.  So, we now update the internal stat struct in the path_item.

Meanwhile, pftool's call to Path::post_process() came *after* the changes
in utime().  In the case of MarFS, this includes a truncate, which changes
the mtime.  Moving the post_process() before the utime() fixes things.

Now, after copying, samefile() will see src and dest as the same file
(provided they are smaller than the default chunksize of 10G), so a restart
with '-n' will correctly skip copying the src to dest.