Don't unlink failed writes.
As a prototype effort to avoid producing imperfectly-written files, we were
unlinking the destination, if the close failed.  First, we fixed a problem
with close, but, second, this was not the appropriate thing to do, anyhow.
In the case of N-to-1 files, there may be many tasks calling close on their
parts, and some of them may fail.

The goal of preventing readers from trying to access such
imperfectly-written files is now handled by installing marfs_restart xattr
at open (as we always did), and removing it in batch_post_process() (in
libmarfs), in the case of successful N-to-1 writes.  Meanwhile, MarFS
internals prevent readers from accessing such files.

pftool's other restart mechanisms can then be used to attempt to fill in
any missing gaps in an N-to-1 files, without worrying that readers may try
to access the file.