Add darshan support as LD_PRELOAD in pfcp, pfls, pfcm
This commit adds the ability to specify darshanlib in the config
file and link in darshan at runtime.  If the darshanlib variable
points at something invalid, then nothing happens.  If
darshanlib points at a valid libdarshan.so (or any other valid
*.so), that file will be linked in via LD_PRELOAD.  This isn't
generally unsafe (the user can use LD_PRELOAD as well), but it
could lead to weird behavior if someone changes the darshanlib
variable point at something ridiculous.

By default, we disable this in the config file, but darshan support
is pretty inocuous, and I expect it is useful in production.