Tweaked read-retries.  Added write-retries under certain conditions.
Read-retries were working, but if there was an error during a close(),
before a successful reopen/reread, pftool was not seeing it as a success.

For MarFS, write-retries only make sense in the limited case where we are
at an object-boundary.  For now, the only time we are sure that is the case
is when copy_file is doing its first write.  (This will also apply for all
the N in an N:1 write.)  In that case, we now support write-retries.

In both cases, the retry operations involve close/open/retry, of the
respective src/dest.  This is probably overkill for POSIX.