The changes in the master branch which converted "work_node" to "out_node" in worker_update_chunk() were committed 2015-08-25 (0ee9fe0). This had the effect of moving the place where CTM xattrs were stored from the source-file to the dest-file.
Our merge in the "cpp" branch got the other parts of the commit, but not
that part. Fixing this now causes us (in the case where destination is not
RO) to "correctly" manage CTM xattrs on the destination, and N:1 restarts
with 'pftool -n' are now working.

De-activated our hack that attempted to work-around the problem in the case
where source-files were read-only.  This was also causing multiple calls to
post_process().