pftool only call pre_process() on first-time destinations.
This means files that have never been written (or which have been unlinked
for re-writing), but not N:1 files which are being restarted with CTM info.

This allows MARFS_Path::pre_process() to avoid having to worry about
truncating N:1 files that are being restarted.  If pre_process() is ever
called on an existing file, there should be error (non-fatal) messages from
pftool, and the file skipped.