non-root user doesn't chmod/chown dest, so MD compare (for non-root) should also skip those.
After recent changes, copying as a non-root user no longer does chmod/chown
on the destination.  So, when using '-n' to see whether a copy task needs
to copy to the destination, or whether the destination "matches" the
source, a non-root user should also not compare the mode, uid, or gid of
the destination to determine whether it matches.