Handle failed obj-overwrites due to out-of-date file-CTM data.
Because it is only updated periodically, file-CTM might not have a full
record of all chunks that have been written.  During restart those chunks
not recorded in the file-CTM will be re-written.  For MarFS destinations
based on recent Scality sproxyd implementations, this will get HTTP-500
errors.  (Scality improves sproxyd performance by making assumptions that
forbid overwrites.)

We now detect this situation.  If the would-be object already exists and
has the size we were going to write to it, we deem this error-situation a
successful write.  (Pftool will already have done high-level validation of
src/dest metadata.)  However, we return a special success-code, so the
output can indicate that the situation occurred.