More like 'cp' behavior, regarding mode-bits and ownership.
By default, always try to chmod the destination to match the source.
This POSIX copy behavior.

For the root-user, or if the new '-o' cmd-line flag is specified, we
also try to chown the destination to match the user/group of the
source.

Failures are considered non-fatal, but they do get reported.

TBD: libmarfs doesn't yet handle umask.