compare-task shouldn't touch destination.
In the case of a "deep" chunked compare, with a MarFS destination,
worker_update_chunk() was writing chunk-info into MarFS MD files.  This
changed the mtime, which caused MD-level comparisons to fail for all
subsequent chunks.  We now only update MarFS MD in COPY tasks.

update_stats() was also getting called at the end of all tasks.  This was
setting MD on the destination to match the source (hiding the previous
problem with MarFS destination MD).  The COMPARE task now skips this.

The result is that "deep" chunked compare now works for MarFS destinations.
We preserved the use of CTM during deep compares, so deep chunked compares
can be restarted, and will pick up where they left off (modulo the fact
that CTM is not updated after every chunk).