Closes Issue 44
The default behaviour was to use ../bin/pftool as the location for pftool. Now it uses an environment variable (PFTOOL) as the location, and failing that, uses the location in the PATH variable, and failing that goes back to the default behaviour. It also lookis for PFTOOL_CONFIG in the environment variables, but falls back to its old default behaviour of ../etc/pftool.cfg