Moved Path::rename_to_original() to Path::rename()
More generic.  Want to preserve the idea that Paths represent a sort
of operating-handle on a particular path.  But a renamed path might
potentially require a different subclass.  No way to assure that
caller would pay attention if we return the new subclass, so instead
we do the rename, but retain the original path, and just invalidate
the stat-info for the old-name, inside the current (correct) Path
sub-class.  Callers can use Path::path_append(), and
Path::path_truncate(), to create new Path objects for the renamed
path.