Long-standing subtle bug in get_base_path() seems unlikely to have been triggered.
Just noticed in passing.  This would've caused the comparison of
dir_name against "." to always fail, but this would be hidden by the
next test unless there were more than one source (or a list) on the
command-line.  In that case, the bug would've caused the base_path for
a directory named './mydir' to be './', rather than './mydir'.