Overhaul of process_stat_buffer()
This function had just gotten too unwieldy.  It was possible to follow
a specific case through the code, but reasoning about general behavior
was difficult.  The problem was flag-variables that were set and
tested in diverse scopes. The recent addition of new special-cases for
temp-file renaming, plus experience trying to chase a production bug,
convinced me it was time.

The new code separates behavior into 3 distinct phases: (1) identify
which files are present, (2) figure out what we are going to do, (3)
do it.